+++ features/admin.applications.v1/constants/application-management.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -376,7 +376,11 @@ export class ApplicationManagementConstants {
     public static readonly TOTP_HANDLERS: string[] = [
         ...ApplicationManagementConstants.FIRST_FACTOR_AUTHENTICATORS,
         ...ApplicationManagementConstants.SOCIAL_AUTHENTICATORS,
-        IdentityProviderManagementConstants.MAGIC_LINK_AUTHENTICATOR
+        IdentityProviderManagementConstants.MAGIC_LINK_AUTHENTICATOR,
+        IdentityProviderManagementConstants.OIDC_AUTHENTICATOR_ID,
+        IdentityProviderManagementConstants.OIDC_AUTHENTICATOR_NAME,
+        IdentityProviderManagementConstants.SAML_AUTHENTICATOR_ID,
+        IdentityProviderManagementConstants.SAML_AUTHENTICATOR_NAME
     ];
 
     // Authenticators that can handle Email OTP.